Awakened gem drop rates now precalculated
Unless I'm mistaken the previous calculated drop rate of exceptional awakened gems was too low. This was because we added all gems in the same pool then divided their drop rates by the amount of gems. This effectively made the drop rate for a specific exceptional drop rates 0.5% / 41, rather than 0.5% / 3.